Moved code around in menu.lua in preparation for encapsulation.
